pyproject.toml
[tool.poetry]
name = "gtm-automation-linkedin"
version = "0.1.0"
description = "Instantly to LinkedIn automation"
authors = ["Alex Puga"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.58.0"
httpx = "^0.28.1"
python-dotenv = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
